requirements.txt
pyTelegramBotAPI==4.12.0
requests==2.31.0
moviepy==1.0.3
pytube==15.0.0
python-dotenv==1.0.0
